StockController finished, double to decimal in StockHistoryRow, little refactoring and bugfixes

+++ Allocator/Allocator/Models/StockHistoryRow.cs
@@ -4,10 +4,10 @@ namespace Allocator.API.Models;
 public class StockHistoryRow
 {
     public int StockHistoryRowId { get; set; }
-    public double Input { get; set; }
+    public decimal Input { get; set; }
     public  DateTime Date { get; set; }
-    public double Profit { get; set; }
-    public double Output { get; set; }
+    public decimal Profit { get; set; }
+    public decimal Output { get; set; }
 
     public int StockId { get; set; }
     public Stock Stock { get; set; }
